restructure api.py; ERROR: GPU OOM

+++ dmff/settings.py
@@ -1,8 +1,9 @@
 from jax.config import config
 
-PRECISION = 'double'  # 'double'
+PRECISION = 'single'  # 'double'
 
 DO_JIT = True
 
 if PRECISION == 'double':
     config.update("jax_enable_x64", True)
+
